Update per review feedback
Fixed a typo in the comment
Put pendingQueue in a separate file
Made T3 timer not to stop (rtx forever)
Other minor cleanups
Resolves #11
